pyproject.toml
[project]
name = "testtaskchestniyznak"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
